LPS-147400 Move common styles from CollectionStylesPanel to CollectionGeneralPanel
