+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/CollectionGeneralPanel.js
@@ -27,6 +27,7 @@ import React, {useCallback, useEffect, useMemo, useState} from 'react';
 
 import {COLLECTION_APPLIED_FILTERS_FRAGMENT_ENTRY_KEY} from '../../../../../../app/config/constants/collectionAppliedFiltersFragmentKey';
 import {COLLECTION_FILTER_FRAGMENT_ENTRY_KEY} from '../../../../../../app/config/constants/collectionFilterFragmentEntryKey';
+import {COMMON_STYLES_ROLES} from '../../../../../../app/config/constants/commonStylesRoles';
 import {FREEMARKER_FRAGMENT_ENTRY_PROCESSOR} from '../../../../../../app/config/constants/freemarkerFragmentEntryProcessor';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../../../../../app/config/constants/layoutDataItemTypes';
 import {config} from '../../../../../../app/config/index';
@@ -40,10 +41,12 @@ import CollectionService from '../../../../../../app/services/CollectionService'
 import InfoItemService from '../../../../../../app/services/InfoItemService';
 import updateCollectionDisplayCollection from '../../../../../../app/thunks/updateCollectionDisplayCollection';
 import updateItemConfig from '../../../../../../app/thunks/updateItemConfig';
+import {getResponsiveConfig} from '../../../../../../app/utils/getResponsiveConfig';
 import {useId} from '../../../../../../app/utils/useId';
 import CollectionSelector from '../../../../../../common/components/CollectionSelector';
 import useControlledState from '../../../../../../core/hooks/useControlledState';
 import CollectionFilterConfigurationModal from '../CollectionFilterConfigurationModal';
+import {CommonStyles} from './CommonStyles';
 
 const LAYOUT_OPTIONS = [
 	{label: Liferay.Language.get('full-width'), value: '1'},
@@ -130,6 +133,15 @@ export function CollectionGeneralPanel({item}) {
 				  ]
 				: [],
 		[collectionConfiguration, setFilterConfigurationVisible]
+	);
+
+	const selectedViewportSize = useSelector(
+		(state) => state.selectedViewportSize
+	);
+
+	const collectionConfig = getResponsiveConfig(
+		item.config,
+		selectedViewportSize
 	);
 
 	const handleCollectionListItemStyleChanged = ({target}) => {
@@ -277,129 +289,144 @@ export function CollectionGeneralPanel({item}) {
 
 	return (
 		<>
-			<CollectionSelector
-				collectionItem={collection}
-				itemSelectorURL={config.collectionSelectorURL}
-				label={Liferay.Language.get('collection')}
-				onCollectionSelect={handleCollectionSelect}
-				optionsMenuItems={optionsMenuItems}
-				shouldPreventCollectionSelect={shouldPreventCollectionSelect}
+			<CommonStyles
+				commonStylesValues={collectionConfig.styles}
+				item={item}
+				role={COMMON_STYLES_ROLES.general}
 			/>
-			{collection && (
-				<>
-					<ClayForm.Group small>
-						<label htmlFor={listStyleId}>
-							{Liferay.Language.get('list-style')}
-						</label>
 
-						<ClaySelectWithOption
-							aria-label={Liferay.Language.get('list-style')}
-							id={listStyleId}
-							onChange={(event) =>
-								handleConfigurationChanged({
-									listStyle: event.target.value,
-								})
-							}
-							options={availableListStyles}
-							value={listStyle}
-						/>
-					</ClayForm.Group>
+			<div className="page-editor__item-general-configuration">
+				<CollectionSelector
+					collectionItem={collection}
+					itemSelectorURL={config.collectionSelectorURL}
+					label={Liferay.Language.get('collection')}
+					onCollectionSelect={handleCollectionSelect}
+					optionsMenuItems={optionsMenuItems}
+					shouldPreventCollectionSelect={
+						shouldPreventCollectionSelect
+					}
+				/>
 
-					{listStyle === LIST_STYLE_GRID && (
+				{collection && (
+					<>
 						<ClayForm.Group small>
-							<label htmlFor={collectionLayoutId}>
-								{Liferay.Language.get('layout')}
+							<label htmlFor={listStyleId}>
+								{Liferay.Language.get('list-style')}
 							</label>
 
 							<ClaySelectWithOption
-								aria-label={Liferay.Language.get('layout')}
-								id={collectionLayoutId}
+								aria-label={Liferay.Language.get('list-style')}
+								id={listStyleId}
 								onChange={(event) =>
 									handleConfigurationChanged({
-										numberOfColumns: event.target.value,
+										listStyle: event.target.value,
 									})
 								}
-								options={LAYOUT_OPTIONS}
-								value={numberOfColumns}
+								options={availableListStyles}
+								value={listStyle}
 							/>
 						</ClayForm.Group>
-					)}
 
-					{listStyle !== LIST_STYLE_GRID &&
-						availableListItemStyles.length > 0 && (
+						{listStyle === LIST_STYLE_GRID && (
 							<ClayForm.Group small>
-								<label htmlFor={collectionListItemStyleId}>
-									{Liferay.Language.get('list-item-style')}
+								<label htmlFor={collectionLayoutId}>
+									{Liferay.Language.get('layout')}
 								</label>
 
-								<ClaySelect
-									aria-label={Liferay.Language.get(
-										'list-item-style'
-									)}
-									id={collectionListItemStyleId}
-									onChange={
-										handleCollectionListItemStyleChanged
+								<ClaySelectWithOption
+									aria-label={Liferay.Language.get('layout')}
+									id={collectionLayoutId}
+									onChange={(event) =>
+										handleConfigurationChanged({
+											numberOfColumns: event.target.value,
+										})
 									}
-								>
-									<ListItemStylesOptions
-										item={item}
-										listItemStyles={availableListItemStyles}
-									/>
-								</ClaySelect>
+									options={LAYOUT_OPTIONS}
+									value={numberOfColumns}
+								/>
 							</ClayForm.Group>
 						)}
 
-					<ClayForm.Group small>
-						<label htmlFor={collectionPaginationTypeId}>
-							{Liferay.Language.get('pagination')}
-						</label>
+						{listStyle !== LIST_STYLE_GRID &&
+							availableListItemStyles.length > 0 && (
+								<ClayForm.Group small>
+									<label htmlFor={collectionListItemStyleId}>
+										{Liferay.Language.get(
+											'list-item-style'
+										)}
+									</label>
 
-						<ClaySelectWithOption
-							aria-label={Liferay.Language.get('pagination')}
-							id={collectionPaginationTypeId}
-							onChange={(event) =>
-								handleConfigurationChanged({
-									paginationType: event.target.value,
-								})
-							}
-							options={PAGINATION_TYPE_OPTIONS}
-							value={paginationType || ''}
-						/>
-					</ClayForm.Group>
+									<ClaySelect
+										aria-label={Liferay.Language.get(
+											'list-item-style'
+										)}
+										id={collectionListItemStyleId}
+										onChange={
+											handleCollectionListItemStyleChanged
+										}
+									>
+										<ListItemStylesOptions
+											item={item}
+											listItemStyles={
+												availableListItemStyles
+											}
+										/>
+									</ClaySelect>
+								</ClayForm.Group>
+							)}
 
-					{paginationType ? (
-						<PaginationOptions
-							displayAllPages={displayAllPages}
-							handleConfigurationChanged={
-								handleConfigurationChanged
-							}
-							initialNumberOfItemsPerPage={
-								initialNumberOfItemsPerPage
-							}
-							initialNumberOfPages={initialNumberOfPages}
-						/>
-					) : (
-						<NoPaginationOptions
-							displayAllItems={displayAllItems}
-							handleConfigurationChanged={
-								handleConfigurationChanged
-							}
-							initialNumberOfItems={initialNumberOfItems}
-							totalNumberOfItems={totalNumberOfItems}
-						/>
-					)}
-				</>
-			)}
+						<ClayForm.Group small>
+							<label htmlFor={collectionPaginationTypeId}>
+								{Liferay.Language.get('pagination')}
+							</label>
 
-			{filterConfigurationVisible ? (
-				<CollectionFilterConfigurationModal
-					collectionConfiguration={collectionConfiguration}
-					handleConfigurationChanged={handleConfigurationChanged}
-					itemConfig={item.config}
-					observer={filterConfigurationObserver}
-					onClose={onFilterConfigurationClose}
-				/>
-			) : null}
+							<ClaySelectWithOption
+								aria-label={Liferay.Language.get('pagination')}
+								id={collectionPaginationTypeId}
+								onChange={(event) =>
+									handleConfigurationChanged({
+										paginationType: event.target.value,
+									})
+								}
+								options={PAGINATION_TYPE_OPTIONS}
+								value={paginationType || ''}
+							/>
+						</ClayForm.Group>
+
+						{paginationType ? (
+							<PaginationOptions
+								displayAllPages={displayAllPages}
+								handleConfigurationChanged={
+									handleConfigurationChanged
+								}
+								initialNumberOfItemsPerPage={
+									initialNumberOfItemsPerPage
+								}
+								initialNumberOfPages={initialNumberOfPages}
+							/>
+						) : (
+							<NoPaginationOptions
+								displayAllItems={displayAllItems}
+								handleConfigurationChanged={
+									handleConfigurationChanged
+								}
+								initialNumberOfItems={initialNumberOfItems}
+								totalNumberOfItems={totalNumberOfItems}
+							/>
+						)}
+					</>
+				)}
+
+				{filterConfigurationVisible ? (
+					<CollectionFilterConfigurationModal
+						collectionConfiguration={collectionConfiguration}
+						handleConfigurationChanged={handleConfigurationChanged}
+						itemConfig={item.config}
+						observer={filterConfigurationObserver}
+						onClose={onFilterConfigurationClose}
+					/>
+				) : null}
+			</div>
 		</>
 	);
 }
