--- modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/CollectionStylesPanel.js
@@ -1,45 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-import PropTypes from 'prop-types';
-import React from 'react';
-
-import {useSelector} from '../../../../../../app/contexts/StoreContext';
-import {getResponsiveConfig} from '../../../../../../app/utils/getResponsiveConfig';
-import {getLayoutDataItemPropTypes} from '../../../../../../prop-types/index';
-import {CommonStyles} from './CommonStyles';
-
-export function CollectionStylesPanel({item}) {
-	const selectedViewportSize = useSelector(
-		(state) => state.selectedViewportSize
-	);
-
-	const collectionConfig = getResponsiveConfig(
-		item.config,
-		selectedViewportSize
-	);
-
-	return (
-		<CommonStyles
-			commonStylesValues={collectionConfig.styles}
-			item={item}
-		/>
-	);
-}
-
-CollectionStylesPanel.propTypes = {
-	item: getLayoutDataItemPropTypes({
-		config: PropTypes.shape({numberOfColumns: PropTypes.number}),
-	}),
-};
